Update PR branches with base before merging
Previously, if the base branch of a PR was updated while the PR was
open, the action would still attempt a merge (as long as all of the PR checks
pass).

This PR introduces a change that checks if the PR-to-be-merged is up to
date with the base. If it isn't, it will trigger an update to merge in
changes from the base (equivalent to the "Update this branch with
master" button in the UI) and not attempt the merge. This update will
trigger a new workflow run, which will eventually cause the merge action
to run again on what should now be an up-to-date PR branch.